requirements.txt
absl-py==0.1.8
bleach==1.5.0
cffi==1.11.2
cycler==0.10.0
enum34==1.1.6
future==0.16.0
h5py==2.7.1
html5lib==0.9999999
Keras==2.1.2
Markdown==2.6.11
matplotlib==2.1.1
numpy==1.14.0
Pillow==5.0.0
protobuf==3.5.1
pycparser==2.18
pyparsing==2.2.0
python-dateutil==2.6.1
pytz==2017.3
PyYAML==3.12
scipy==1.0.0
six==1.11.0
tensorflow==1.5.0rc0
tensorflow-tensorboard==0.4.0rc3
Werkzeug==0.14.1
